Allow subclasses to manipulate reachabilityFlags
Previously, many methods made a direct call to
SCNetworkReachabilityGetFlags(reachabilityRef, &flags) to get the flags.
Since we already have a method that does this, we can "program to an
interface" and have all flag queries go through `-reachabilityFlags` to
gain readability (IMO) and more importantly the ability of subclasses
overriding `-reachabilityFlags` for testing purposes.